{{cookiecutter.project_directory}}/pyproject.toml
[project]
name = "cookiecutter.project_directory"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "GGGoingdown", email = "gggoingdown0401@gmail.com"},
]
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "httpx>=0.26.0",
    "sentry-sdk>=1.40.4",
    "tortoise-orm[asyncpg]>=0.20.0",
    "tenacity>=8.2.3",
    "loguru>=0.7.2",
    "aioredis>=2.0.1",
    "pydantic-settings>=2.1.0",
    "aerich>=0.7.2",
    "dependency-injector>=4.41.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"

[tool.pdm.dev-dependencies]
dev = [
    "httpx>=0.26.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
]

[tool.aerich]
tortoise_orm = "src.pkg.db.DEFAULT_TORTOISE"
location = "./migrations"
src_folder = "./."
